Add a customized SnalLatch design
The SnapLatch design in the VoronUsers repository is designed to
use M3x12 screws. I wanted to be able to use M3x8. While the
original model did have a user parameter for the length of the
screw, it did not allow for the specification of the screw stickout.
Rather, it had the stickout hard coded to 5mm.

This seemed to leave too little material in the hinge body. To
solve this, I added another user parameter that allows the user to
customize the screw stickout. The default is set to 4mm.