Add custom idle timeout handling
The default idle timeout handling was defined to turn off all the
heaters and disable motors. The act of disabling the motors caused
the printer to lose its position.

As a result, if the idle timeout was hit while the printer was
paused, the print could not be resumed (since homing and QGL is not
possible).

This commit adds a custom idle timeout handling, which checks if the
printer is an paused state. If so, it just turns off the extruder
heater.