Add macro for changing filament
Changing filament requires heating up the hotend to the extrusion
temp of the currently loaded filament, unloading it, adjusting the
extruder temp to the temperature of the new filament, and then
loading the new one.

The LOAD_FILAMENT macro will extrude 100mm in order to purge the
old filament.

This macro partially automates and makes the process more convinient
by moving the head to a front of the printer, high above the bed,
and heating up the extruder to the desired temperature.

This makes the swap and removal of the purged filament easier.