Add custom print control menu options
The default display menu had some print control options but those
were geared towards OctoPrint users and did not work with Mainsail.

Disable the OctoPrint menu and add a custom one that contains the
same items.

("index" seems to be a hidden/undocumented option for menu items
that allows the user to order the menu items in the desired order.)