Add the BTT Smart Filament Sensor to the printer
This change adds the configuration for the BTT Smart Filament
Sensor. It also updates PRINTER_INIT, PRINT_START, and PRINT_END
to disable the sensor by default and only enable it during printing.

For now, the actions performed during filament runout detection is
just a simple PAUSE. This should be improved to be more useful.