Allow filament to ooze at print end
This is an experimental change to try to better handle filament
oozing.

The idea is to keep the extruder heater on for an additional 60
seconds after the print has ended to allow the filament to ooze
out. Then the extra filament can be cleaned before the
next print starts rather than during extruder heat-up.