Add per-surface Z offset controls
This change add support for Z offset control on a per-surface
level.

Different print surface will require different Z offsets to get
good first layers. With this change, the Z offsets can be configured
and saved for future use.

Once configured, the surface to be used can be set from the display
or through a macro. This will set the GCode offset for that
surface.