Rework how pressure advance is set
This commit reworks the method for setting pressure advance settings.

With this change, the SET_PRESSURE_ADVANCE macro has been
redefined to take two additional parameters - FILAMENT_ID and
FILAMENT_TYPE.The default arguments (ADVANCE and
SMOOTH_TIME) are still accepted and they take presedence.
However, when not specified, the SET_PRESSURE_ADVANCE macro
uses the FILAMENT_ID and FILAMENT_TYPE arguments to look up
the pressure advance values from the USER_VARIABLES.

The pressure advance settings used to be set in the PRINT_START
macro. This has been removed and the printer profiles have been
changed to include a direct call to SET_PRESSURE_ADVANCE in the
start GCode.