Remove debugging and unused code
Remove the debugging and unused code.

This change also fixes a crash with current version of Klipper caused
by calling reactor.unregister_timer(). It seems that the reactor is
not finding our timer, even though it was registered with
reactor.register_timer() and we are passing the returned
timer handle.