Add per-filament Z-offset to PRINT_START
When tuning PolyMaker PETG, it turns out that it requires a different
Z offset in order to get a good first layer.

This commit adds the ability to apply a per-filament relative Z offset
during PRINT_START. The offset gets applied after the last homing
and prior to handing control off to the slicer GCode.

The offset is reset in PRINT_END.

Per-filament Z offsets will come from the USER_VARIABLES macro.