Initial implementation of adaptive bedmesh
In order to save time on performing the bedmesh calibration, adaptive
bedmesh computes the area used by the print and does a bedmesh
calibration only on that area.

Something like this already exists using pure macros
(https://github.com/Frix-x/klipper-voron-V2/blob/main/macros/probing/bed_mesh.cfg).
However, it relies on configuration in the slicer. This
implementation parses the GCode file using the same module
that Moonraker uses to parse for object exclusion. The benefit
of this is that it does not require any slicer configuration and all
slicers supported by the Moonraker module are supported by
this implementation.