Add support for file exclusions to update.py tool
Allow for exclusion of files based on shell file patterns. Files matching
any of the patterns will be excluded from all operations.

This commit also adds an exclusion file containing the Obico config
file. The file contains a server access token, which should not be
exposed. Therefore, the file can not be checked into GitHub.