Allow preservation of probe point spacing
By default, the spacing between the probe points in the default bed
mesh was static in both X and Y directions. This was because the spacing
was defined by the size of the bed and the number of probe points (which
for a square bed are equal for both X and Y).

When adaptive bedmesh is used, the number of probe points is a
fraction of the default probe count. Because the probe count is
generally a small integer value, this could lead to the spacing between
probe points to increase. This results in low of fidelity.

This commit adds a new option, which can be used to alter the adaptive
bed mesh algorithm: --use-spacing[=<percent>].

When this option is used, the probe point generation algorithm atempts
to preserve the spacing between the probe points. This may result in
the number of probe points in one or both direction to be higher than
when the option is not used.