Add Klicky probe dock/undock test script
This commit adds a script to test the reliability of the Klicky probe's
dock/undock procedure.

The script was heavily inspired by the probe_accuracy_test script
from https://github.com/sporkus/probe_accuracy_tests. It uses
Moonraker's Client API to control the printer. For this reason, this
commit also includes a small Python module that wraps the Moonraker
API in a Python class.