Implement a new class to represent a printer configuration
Dealing with both local and remote printer configuration files was
getting to be painful, especially when multiple sub-directories are
used.

To make this easier, implement a class that represents a printer
configuration. The class can maintain state related to the configuration
that will make dealing with all configuration files, regardless of
layout easier.