Switch webcam configuration to Crowsnest
Crowsnest is a new service implemented by the Mainsail team that uses
ustreamer for webcam streaming. ustream is a much more efficient and,
hense, consumes much less resources.

This commit switches the webcam configuration to Crowsnest. The old
webcam.txt configuration file is not removed for now but it is
included in update-excludes.txt so it is skipped.