Fix relative_reference_index computation in adaptive_bed_mesh.py
The goal of the relative_reference_index (RRI) is to pick a point that
is as close to the center of the meshed area as possible. That point
will be considered the base height for the mesh.

The previous computation of the RRI index was flawed in that it would
end up picking the wrong probe point if the total number of probe points
was even. For example, with probe_count=4,4, the probe points look like:

    0  1  2  3
    4  5  6  7
    8  9  10 11
    12 13 14 15

The RRI index computed would be (16/2) + (16%2) = 8. Index 8 is not in
the middle of mesh.

With the new formula, the RRI index for the same mesh would be
(((16-1)/2)-1) + (16%2) = 6. For meshes with odd number of total probe
points (5x5), the result would be (((25-1)/2)-1)+(25%2) = 12.