Add option to maintain specific spacing between probes
The script already has an option to maintain the ratio of spacing
between probe points. However, for prints that are very imbalanced
in terms of size, this can lead to probe points being dense in one
direction and spaced out in the other.

This commit adds a new command line option to allow the user to
specify the spacing (in mm) between probe points. When used, the
script will adjust the number of probe points for each axis in
order to maintain the spacing.