Change LCD display name
Mainsail 2.4.0 adds support for NeoPixels. This causes the display
and case light LEDs to show up in the control.

Previously, the name used for the LCD display didn't matter since it
was never seen. However, now it's displayed as "Btt Mini12864", which
is not very nice or descriptive.

Since there is only going to be one LCD display ever, change the name
of the display NeoPixels to "lcd_display". This changes how it's
displayed to "Lcd Display".