Pull Klicky from external source
This commit reworks how the Klicky macros are maintained in this repo
and on the printers.

Previously, this repo included modified version of the Klicky macros.
While this worked fine, it was not simple to update to newer versions
from the upstream repo. There was a script that attempted to automate
that but it only allowed for simple changes to be made - changes that
can be done using regular expression substitutions.

Since then, I have forked the Klicky_Probe repo and created a branch in
my fork that contains all of my personal changes. This way, I can pull
changes from the upstream "main" branch and rebase my custom branch on
top.

This commit adds the infrastructure to pull updates configs from the
upstream repo by adding the ability to update this repo from external
sources.

Along with those changes, the Klicky macros have been updated to the
top of the custom branch.