Add custom Klipper probe module that can discard the first sample
For some reason, the first probe sample that Klipper makes seems to be
incorrect. This can affect the offset calculations. Even though,
discarding the first probe sample has been requested as a change, it has
not been implemented.

This commit adds a custom module that adds an option to enable throwing
the first sample away.