Allow using the update script to update other locations
Previously, the update script would only handle files from the
printer/config directory. However, it is helpful to be able to automate
the update of Klipper modules and/or other tools.

This commit enhances the script to allow it to update other directories
based on settings in the configuration file.