Turn bed heater to full power
As it turns out, it is not necessary to limit the maximum power of the
bed heater. The formula of 0.4W/cm^2 is the minimum power required to be
able to bring the bed up to ABS temperature.

There was a misunderstanding that any more than the above would result
in the bed taco'ing. However, it would take much more power to
permanently warp a 6mm bed. So, it's safe to set the max_power parameter
to it's default (1.0)