Turn off part cooling fan as part of PRINT_END
Normally, the part cooling fan is being turned off my the GCode
generated by the slicer. It outputs a M107 command right before it
calls PRINT_END.

However, if the printer is cancelled, that command is not executed,
which leaves the fan running. Fix this by explicitly turning the fan
off in PRINT_END.