Fix bug with filament load/unload macros
The filament load/unload macros lookup the maximum extrusion length
from the "extruder" section in order to run the extruder motor
correctly without causing issues.

However, if the option ("max_extrude_only_distance") is not specified,
the macros resulted in an error. Fix this my providing a default value
that matches the default used by Klipper.