Add the Logitech C270 Camera mount mod
Add the custom Logitech C270 camera mount mod. This camera mount is
meant to be adjustable so a proper/desired viewing angle can be
achieved.

The main reason for developing this mount is so I could mount my
camera on the rear gantry extrusion and be able to position it so the
nozzle is in the middle of the picture.