Moved test/debug macros to their own sub-directory
One of the macros that was in debug.cfg was the TEST_SPEED macro from
Ellis'es tuning guide. However, to allow for easier updates to the
macro, it's better to have it in it's own file. This way it can just be
downloaded from GitHub.

So, remove the macro from debug.cfg and move debug.cfg and the new
TEST_SPEED.cfg into their own debug directory.