Fix a bug when uploading to non-existing directories
The upload_file() method of the PrinterConfig class is supposed to check
if a file or any of the file's parent directories exist. If any of the
parent directories do not exist, the method is supposed to create them
so the file can be uploaded to the correct location.

There was a bug in the algorithm used to create all missing sub-
directores, which lead to the last sub-directory to be omitted. This
caused an exception since the method could not successfully upload the
file.

Fix the algorithm so all parent sub-directories are checked and created
if missing.