Use object parameters from Gcode file in present
The Moonraker object exclusion/cancellation pre-processor checks the
GCode file and if it has already been processed does not redo the work.
This results in the adaptive bed mesh script not getting an object
definitions.

The fix is pretty simple: since the adaptive bed mesh script uses the
exact same object data, check the original GCode file first. If the
data is already in the GCode file, use it. Otherwise, pass the file
through the processor.