Revert QGL probe points to original values
Commit 2107269 ("Define QGL probe points closer to the edge of the bed")
moved the QGL probe points closer to the edges of the bed. The idea was
that it might provide better gantry leveling by extending the bed area
over which the gantry is leveled.

However, it seems that these probe points are proving problematic when
printing with elevated chamber temperatures. At ABS temps, QGL saw
multiple failures with probe tolerance errors.

It is not certain that the cause of the probe failures are the extended
probe points. However, the new points have not provided better leveling
so they are being reverted as a safety precaution.