Consider all object polygon points
(This commit fixes a couple of issues and includes some code re-
formatting/re-structuring. However, one of the fixed issues is big
enough as to give the commit its title.)

The object coordinates generated by the exclude object processor is not
just a simple box that describes each object. It's actually a multi-
point polygon. The logic in adaptive_bed_mesh.py, however, assume that
each object will be described by only 4 points. As a result, it hard-
coded the point indexes into the object polygon.

This resulted in the tool generating a bed mesh area that did not fully
represent the used print area.

This commit changes the logic for determining the active print area by
considering all of the points in each object's polygon.

A second fix in this commit is to parse the passed-in probe counts into
a Point object. The output function expects it to be a Point object.
However, the code path that was used when generating a bed mesh area
based on the original data passed in a tuple.