Complete overhaul of the display and menu setup
This commit completely overhauls the LCD display and menu configurations.
Instead of trying to adjust and insert custom menu items in the default
Klipper menu, this change includes a complete custom menu.

It also completely redefines the home/status screen. The changes are
based on the custom display and menu setups from
https://github.com/zellneralex/klipper_config.