Add ability to compile and flash firmware from tool
This commit adds an !!EXPERIMENTAL!! feature to build and flash the host
and MCU firmware from within the tool.

When this mode is used, the tool will upload the appropriate firmware
configuration file, and based on the operation selected, build and/or
flash the firmware to the appropriate controller.

Please note that while the building of the firmware was tested, the
flashing has not since I don't have a spare set of HW to test on!!