Add support for tracking carbon filter usage
The activated carbon filter material needs to be changed periodically
because it does get saturated diminishing its filtering ability.

This commit adds a set of macros that use the print statistics to keep
track of the service duration of the carbon and display a message if it
needs to be changed.