Alter meshing algorithm based on generated probe points
Rather than trying to change the number of probe points to fit the
requirements of the algorithm, do the opposite - change the algorithm
based on the number of probe points generated.

As a added bonus, alter the number of probe points so there is always
a probe in the middle of the meshed area.