Rename and improve print variables macro
The print variables macro is a "storage" macro used to store variables
related to the current print so they can be used by other macros besides
PRINT_START.

This commit renames the macro to be a bit more descriptive and also
improves how the variables are set. Rather than going through a
variable rename (parameter name to variable name), just match parameter
and variable names. This way the values can be much more easily set
from either the macro parameters or the defaults.