Improve HEAT_SOAK macro
The heat soak macro had an issue where the temperature of the chamber
would drop as soon as the Nevermore speed was reduced. This can't be
fully fixed since calling the HEAT_SOAK macro without a soak delay
would still result in the Nevermore turning down and the temperature
dropping. However, it can be improved if a soak duration is passed in.

The macro computes a set of Nevermore fan speeds that will gradually
drop the fan's speed in even increments across evenly spaced time
intervals. This seems to help with maintaining chamber temperature and
still allowing the fan to reach its exit speed.