Update configuration for the state_notify.py module
Switch to using the state_notify.py module from the
voron-klipper-extensions repository. This module removes the need for
a lot of custom configuration for running gcode when the printer is not
active. It also removes the need for modifications to the idle_timeout
Klipper module, which has the benefit it allowing updates from the UI
to work again.