Tuning and improvements to load/unload macros
The changes in this commit are based on tuning and testing of the load/
unload/filament change macros when attempting to do a dual color print
with a single nozzle.

The macros should be more reliable and filament changes should be easier
to do, especially if a change is needed while the chamber is at temp.