Fix a but with user_input() in update script
The user_input() function generates a selection possibilities list based
on the provided set of options. When the provided options is a dict(),
it uses the dictionary keys are the list. However, dict.keys() returns
a dict_keys object, not a list.

This causes an error since dict_keys does not provide an index() method.
To fix this, turn the dict_keys object into an actual list.