Remove unused settings in Auto-Z Calibration config file
Recently, there was an update to the Auto-Z Calibration module, which
renamed some configuration variables.

While those variables were not being used in my config, it's extra
maintenance to go through and update them if they are not being used.

So, to simplify things, just remove all unused/unnecessary settings and
lines from the config.