Clean the nozzle before Z calibration
Previously, the nozzle was being cleaned prior to QGL and Bed Mesh. This
was done because I wanted to avoid cleaning the nozzle with the probe
docked as to avoid accidentally knocking it off.

However, it's better if the cleaning was done right before Z calibration
since that will avoid any extra oozing during QGL and Bed Mesh. Even
though, this requires one extra dock/unlock sequence, it produces better
results.