Split HW definitions into their own directory
This is an re-organization change that splits all HW definitions and
support macros into a separate HW directory.

The main reason for this split is that the includes can be ordered
better. With this split, all HW definitions are included first. This
causes Klipper to create the internal HW objects prior to anything else
(Klipper creates objects as it processes sections from the
configuration).