Adapt configuration and tools for new bed mesh zero position
Klipper has deprecated the `relative_reference_index` bed mesh setting
and replaced it with `zero_reference_index`. This new setting is "static"
in the sense that it cannot be changed after initial definition on
printer startup. As a result, new RRIs do not have to be computed when
using adaptive bed mesh.

This commit updates all of the configuration for this new change.