Increase safe_z_homing z_hop to 15mm
It seems that at higher bed temperatures, the bed expands enough
to where the z_hop of 10mm is not enough for the height of the
Klicky probe. As a result, the probe drags on the bed and reports
triggered immediately, which results in the error "Probe triggered
prior to movement."

Fix this my increasing the z_hop to 15mm so the toolhead is lifted
a bit more after homing.