Add configuration for the skirt buttons board
This commit adds the configuration for the skirt buttons MCU
(https://github.com/3DAixperts/skirt-button-pcb).

The commit includes Klipper firmware configuration, MCU and button HW
configuration, a set of macros to handle the button presses along with
color changes for various activities, and changes to the update script
to handle updading the MCU firmware.