Allow for a delay before turning off the extruder on PAUSE
In the case of the filament sensor detecting a false positive trigger,
it is helpful to prevent the extruder from being turned off immediately
since that causes the temperature to drop quickly.

Instead, enhance the PAUSE macro to accept new parameters that will
allow control of whether the extruder should be turned off and if so,
whether it should be turned off immediately or after a delay.

This new capability is used in the filament sensor's 'on_runout_gcode'
section.