Use Klipper action instead of GCode for EStop button
Using the M112 GCode as the EStop button gcode does work correctly since
Klipper will execute the M112 GCode as it gets to in the GCode stream.
This can result in the EStop not actually stopping the machine until the
stream gets to it.

Using Klipper actions produce an immediate effect, which is what should
happen.