Update the sync command to allow downloading remote-only files
The sync command did not allow for downloading files that are only on
the printer. This meant that part of the sync operation could not be
done (pulling files from the printer to the local repository).

Fix this by adding support for downloading files. Adding this support
meant fixing some more issues with how user input selection is returned
to the caller.