Fix links in README file
The top level README file was changed to include a table of printer
specs. The table had to be done with pure HTML since GitHub's markdown
does not support multi-row cells.

It seems that GitHub does not process HTML chunks and leaves them as-is.
Therefore, the markdown-style links in the table cells were left as-is.

Fix this by changing them to HTML-style links.