Fix park selection in display menu
The display menu contains a sub-menu that allows for selecting the park
position. However, that did not work because the menu code was trying to
index into a dict_keys() object, which does not support subscripting.

Furthermore, Jinja does not have a way to index into a list created by
using the `list` filter, which is required because the menu configuration
cannot execute multi-line templates for the `name` setting.

So, the solution was to dynamically create a list of park position names
and index into that.