Reposition Z endstop
I am dealing with some issues related to Auto-Z calibration failing when
probing the nozzle tip on the Z endstop. When probing at ABS temperatures,
the endstop seems to give bad results, which errors out the process.

To hellp with this (although, ultimately the switch may have to be
replaced), I've moved the endstop further away from the bed. The thinking
is that it might be getting too hot (since this does not seem to be
happening when printing PLA).