Fix Moonraker update_manager config for two extensions
Moonraker expects that the `virtualenv` parameter in the `[update_manager]`
sections for extension to point to the toplevel directory of the
environment.

This was not the case for two extensions - mobileraker_companion and
moonraker-obico. This commit fixes this and also moves the moonraker-obico
config into the moonraker.conf file (moonraker-obico-update.cfg) is removed.