Redo chamber heat soak fan control
The chamber fan control in the heat soak macro used to slowly ramp up
the fan speed. The goal was to avoid large drops in the bed temp when
the fan kicked in.

However, that is not the recommended speed stepping. One of the side
effects of the wrong control was that the filter cover would warp due
to the high bed temp.

Redo the fan control so it starts off at a high speed and then drops
down as the chamber heats up.